# [perf] additional performance improvements for OTAP encoding