--- a/rust/otap-dataflow/crates/otap/src/encoder/cbor.rs
+++ b/rust/otap-dataflow/crates/otap/src/encoder/cbor.rs
@@ -3,22 +3,93 @@
 
 use std::io::Write;
 
-use ciborium::{Value, value::Integer};
 use otap_df_pdata_views::views::common::{AnyValueView, AttributeView, ValueType};
+use serde::ser::{SerializeMap, SerializeSeq, Serializer};
+use serde_cbor::ser::IoWrite;
 
 use crate::encoder::error::Result;
 
-/// serialize the passed `AnyValue` view as bytes in cbor representation
+/// Adapter for serializing AnyValueView using Serde
+struct AnyValueSerializerWrapper<T>(pub T);
+
+impl<'a, T> serde::Serialize for AnyValueSerializerWrapper<T>
+where
+    T: AnyValueView<'a> + 'a,
+{
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serialize_any_value::<T, S>(&self.0, serializer)
+    }
+}
+
 pub fn serialize_any_values<'a, I, T, W>(source: I, writer: W) -> Result<()>
 where
     I: Iterator<Item = T>,
     T: AnyValueView<'a> + 'a,
     W: Write,
 {
-    let value = Value::Array(source.map(|val| cbor_value_from_any_value(&val)).collect());
-    ciborium::into_writer(&value, writer)?;
+    let mut serializer = serde_cbor::Serializer::new(IoWrite::new(writer));
+    let mut seq = serializer.serialize_seq(None)?;
+
+    for val in source {
+        seq.serialize_element(&AnyValueSerializerWrapper(val))?;
+    }
+    SerializeSeq::end(seq)?;
 
     Ok(())
+}
+
+/// serialize the passed `AnyValue` view as bytes in cbor representation
+fn serialize_any_value<'a, T, S>(source: &T, serializer: S) -> std::result::Result<S::Ok, S::Error>
+where
+    T: AnyValueView<'a> + 'a,
+    S: Serializer,
+{
+    match source.value_type() {
+        ValueType::String => {
+            let s = source.as_string().expect("expected string");
+            serializer.serialize_str(s.as_ref())
+        }
+        ValueType::Bool => {
+            let b = source.as_bool().expect("expected bool");
+            serializer.serialize_bool(b)
+        }
+        ValueType::Int64 => {
+            let i = source.as_int64().expect("expected int64");
+            serializer.serialize_i64(i)
+        }
+        ValueType::Double => {
+            let f = source.as_double().expect("expected f64");
+            serializer.serialize_f64(f)
+        }
+        ValueType::Bytes => {
+            let b = source.as_bytes().expect("expected bytes");
+            serializer.serialize_bytes(b)
+        }
+        ValueType::Empty => serializer.serialize_none(),
+        ValueType::Array => {
+            let children = source.as_array().expect("expected array");
+            let mut seq = serializer.serialize_seq(None)?;
+            for child in children {
+                seq.serialize_element(&AnyValueSerializerWrapper(child))?;
+            }
+            seq.end()
+        }
+        ValueType::KeyValueList => {
+            let kvlist = source.as_kvlist().expect("expected kvlist");
+            let mut map = serializer.serialize_map(None)?;
+            for kv in kvlist {
+                let key = kv.key();
+                match kv.value() {
+                    Some(v) => map.serialize_entry(&key, &AnyValueSerializerWrapper(v))?,
+                    None => map.serialize_entry(&key, &Option::<()>::None)?,
+                }
+            }
+            map.end()
+        }
+    }
 }
 
 /// serialize the passed list of key values as bytes in cbor representation
@@ -28,57 +99,18 @@ where
     T: AttributeView,
     W: Write,
 {
-    let value = cbor_value_from_kvlist(source);
-    ciborium::into_writer(&value, writer)?;
+    let mut serializer = serde_cbor::Serializer::new(IoWrite::new(writer));
+    let mut map = serializer.serialize_map(None)?;
+    for kv in source {
+        let key = kv.key();
+        match kv.value() {
+            Some(v) => map.serialize_entry(&key, &AnyValueSerializerWrapper(v))?,
+            None => map.serialize_entry(&key, &Option::<()>::None)?,
+        }
+    }
+    SerializeMap::end(map)?;
 
     Ok(())
-}
-
-/// convert the any value to cbor `Value`
-fn cbor_value_from_any_value<'a, T>(source: &T) -> Value
-where
-    T: AnyValueView<'a>,
-{
-    match source.value_type() {
-        ValueType::String => Value::Text(source.as_string().expect("body is string").to_string()),
-        ValueType::Bool => Value::Bool(source.as_bool().expect("body is bool")),
-        ValueType::Int64 => Value::Integer(Integer::from(source.as_int64().expect("body is int"))),
-        ValueType::Double => Value::Float(source.as_double().expect("body is float")),
-        ValueType::Bytes => Value::Bytes(source.as_bytes().expect("body is bytes").to_vec()),
-        ValueType::Array => {
-            let children = source
-                .as_array()
-                .expect("body is array")
-                .map(|val| cbor_value_from_any_value(&val))
-                .collect();
-            Value::Array(children)
-        }
-        ValueType::KeyValueList => {
-            cbor_value_from_kvlist(source.as_kvlist().expect("body is kvlist"))
-        }
-        ValueType::Empty => Value::Null,
-    }
-}
-
-/// convert the key-value pairs into cbor `Value`
-fn cbor_value_from_kvlist<I, T>(source: I) -> Value
-where
-    I: Iterator<Item = T>,
-    T: AttributeView,
-{
-    let children = source
-        .map(|kv| {
-            (
-                Value::Text(kv.key().to_string()),
-                match kv.value() {
-                    None => Value::Null,
-                    Some(val) => cbor_value_from_any_value(&val),
-                },
-            )
-        })
-        .collect();
-
-    Value::Map(children)
 }
 
 #[cfg(test)]