--- a/rust/otel-arrow-rust/src/encode/record/array/binary.rs
+++ b/rust/otel-arrow-rust/src/encode/record/array/binary.rs
@@ -22,16 +22,32 @@ use super::{ArrayBuilder, ArrayBuilderConstructor};
 impl ArrayAppend for BinaryBuilder {
     type Native = Vec<u8>;
 
+    #[inline(always)]
     fn append_value(&mut self, value: &Self::Native) {
         self.append_value(value);
+    }
+
+    #[inline(always)]
+    fn append_values(&mut self, value: &Self::Native, n: usize) {
+        for _ in 0..n {
+            self.append_value(value);
+        }
     }
 }
 
 impl ArrayAppendSlice for BinaryBuilder {
     type Native = u8;
 
+    #[inline(always)]
     fn append_slice(&mut self, value: &[Self::Native]) {
         self.append_value(value);
+    }
+
+    #[inline(always)]
+    fn append_slice_n(&mut self, value: &[Self::Native], n: usize) {
+        for _ in 0..n {
+            self.append_value(value);
+        }
     }
 }
 
@@ -41,12 +57,7 @@ impl ArrayAppendNulls for BinaryBuilder {
     }
 
     fn append_nulls(&mut self, n: usize) {
-        // TODO - to be more efficient, after the next release of arrow-rs we should revisit this
-        // and call append_nulls on the base builder. Waiting on these changes to be released:
-        // https://github.com/apache/arrow-rs/pull/7606
-        for _ in 0..n {
-            self.append_null();
-        }
+        self.append_nulls(n);
     }
 }
 
@@ -91,6 +102,10 @@ where
     fn append_value(&mut self, value: &Self::Native) -> dictionary::Result<usize> {
         self.append_slice(value)
     }
+
+    fn append_values(&mut self, value: &Self::Native, n: usize) -> dictionary::Result<usize> {
+        self.append_slice_n(value, n)
+    }
 }
 
 impl<K> DictionaryArrayAppendSlice for BinaryDictionaryBuilder<K>
@@ -102,6 +117,18 @@ where
 
     fn append_slice(&mut self, value: &[Self::Native]) -> dictionary::Result<usize> {
         match self.append(value) {
+            Ok(index) => Ok(index.into()),
+            Err(arrow::error::ArrowError::DictionaryKeyOverflowError) => {
+                Err(dictionary::DictionaryBuilderError::DictOverflow {})
+            }
+            // safety: shouldn't happen. The only error type append should
+            // return should be for dictionary overflows
+            Err(e) => panic!("unexpected error type appending to dictionary {e}"),
+        }
+    }
+
+    fn append_slice_n(&mut self, value: &[Self::Native], n: usize) -> dictionary::Result<usize> {
+        match self.append_n(value, n) {
             Ok(index) => Ok(index.into()),
             Err(arrow::error::ArrowError::DictionaryKeyOverflowError) => {
                 Err(dictionary::DictionaryBuilderError::DictOverflow {})