--- a/rust/otel-arrow-rust/src/encode/record/array/boolean.rs
+++ b/rust/otel-arrow-rust/src/encode/record/array/boolean.rs
@@ -41,11 +41,6 @@ impl AdaptiveBooleanArrayBuilder {
 
     pub fn append_value(&mut self, value: bool) {
         if self.inner.is_none() {
-            // TODO -- when we handle nulls here we need to keep track of how many
-            // nulls have been appended before the first value, and prefix this
-            // newly initialized array with that number of nulls
-            // https://github.com/open-telemetry/otel-arrow/issues/534
-
             self.inner = Some(BooleanBuilder::new());
             if self.nulls_prefix > 0 {
                 self.append_nulls(self.nulls_prefix);