--- a/rust/otel-arrow-rust/src/encode/record/array/primitive.rs
+++ b/rust/otel-arrow-rust/src/encode/record/array/primitive.rs
@@ -28,7 +28,7 @@ where
         self.append_value(*value);
     }
 
-    fn append_values(&mut self, value: &Self::Native, n: usize) {
+    fn append_value_n(&mut self, value: &Self::Native, n: usize) {
         self.append_value_n(*value, n);
     }
 }