… rust/otap-dataflow/Cargo.toml
… rust/otap-dataflow/benchmarks/benches/otap_encoder/main.rs
… rust/otap-dataflow/crates/otap/Cargo.toml
… rust/otap-dataflow/deny.toml
… rust/otel-arrow-rust/Cargo.toml
… rust/otel-arrow-rust/src/encode/record/array/fixed_size_binary.rs
… rust/otel-arrow-rust/src/encode/record/array/string.rs
… rust/otel-arrow-rust/src/encode/record/logs.rs