### Changes:

**Benchmark improvements:**
- Adds a variety of batch sizes to the `otap_encoder` benchmark

**Performance improvements**
- Adaptive Array Builders:
  - Adds new methods to append multiple values to the array builders in a single method call. This is 
  - Use the optimized `append_nulls` implementation that was introduced in arrow-rs v52.2 https://github.com/apache/arrow-rs/pull/7606
  - Uses a macro to handle the append value logic instead of a method with callbacks instead of the generic `handle_append` method that took callback arguments. This improves performance on the hot path of `AdaptiveArrayBuilder::append_value / append_slice / etc.`
- Add additional inlines for some short, frequently called methods
- Array/KvList value serialization:
  - optimize CBOR serialization to remove heap allocations (previously we were allocating `vec` in several places when creating `ciborium::Value`, but now we use `serde_cbor::Serializer` to avoid the allocations)
  
### Benchmark Results:

Comparing benchmark **`encode_otap_logs_using_views/proto_bytes->views->OTAP`** on old code vs new. 

| batch size | old code | new code | % diff |
| --- | --- | --- | --- |
| resources=5,scopes=10,logs=5 | 23.571 us | 19.568 us | 16.98% |
| resources=1,scopes=1,logs=1000 | 74.841us | 61.622us | 17.66% |
| resources=1,scopes=1,logs=5000 | 3.6978ms | 3.0042ms | 18.75% |
| resources=1,scopes=1,logs=5000 | 7.3788ms | 6.0012ms | 18.67% |


Comparing **`encode_otap_logs_using_views/proto_bytes->views->OTAP`** with **`decode_proto_bytes/proto_bytes->prost`**.

| batch size | proto_bytes->prost | proto_bytes->views->OTAP | % diff |
| --- | --- | --- | --- |
| resources=5,scopes=10,logs=5 | 15.215us | 19.568 us | 28.61% |
| resources=1,scopes=1,logs=1000 | 55.204us | 61.622us | 11.63% |
| resources=1,scopes=1,logs=5000 | 2.7839ms | 3.0042ms | 7.91% |
| resources=1,scopes=1,logs=5000 | 5.5983ms | 6.0012ms | 7.20% |

Benchmark machine:
```
macOS 15.5 (24F74)
CPU: Apple M4 Pro
Memory: 24 GB
```

---

resolves: https://github.com/open-telemetry/otel-arrow/issues/534